Removed old realisations of Header component.
