--- src/components/Header/Header.scss
@@ -1,275 +1,0 @@
-@import '@styles/variables/all';
-@import '@styles/mixins/all';
-
-header {
-  -webkit-box-sizing: border-box;
-  -moz-box-sizing: border-box;
-  -ms-box-sizing: border-box;
-  box-sizing: border-box;
-  padding: 2.5rem 4.28rem 0;
-
-  @include media-breakpoint-down(sm) {
-    padding: 2rem 0 0;
-  }
-}
-.main-header {
-  position: fixed;
-  height: 10rem;
-  top: 0;
-  left: 0;
-  z-index: 11;
-  width: 100%;
-  background: linear-gradient(0deg, rgba(2, 2, 30, 0.0001) 0%, #02021e 98%);
-
-  @include media-breakpoint-down(md) {
-    padding: 2.5rem 5rem 0;
-  }
-
-  @include media-breakpoint-down(xs) {
-    padding: 2rem 1.67rem 0;
-    height: 6.67rem;
-  }
-}
-
-header .logo {
-  height: auto;
-  max-width: 100%;
-  width: 9.5rem;
-  @include media-breakpoint-down(sm) {
-    width: 10rem;
-  }
-}
-header .mail {
-  color: #fff;
-  font-family: $font-sofia-pro-soft;
-  font-size: 1.6rem;
-  font-weight: 300;
-  line-height: 2.8rem;
-  position: relative;
-  text-decoration: none;
-  -webkit-transition: 0.2s;
-  -o-transition: 0.2s;
-  transition: 0.2s;
-  @include media-breakpoint-down(sm) {
-    font-size: 2rem;
-    .desctop_mail {
-      display: none;
-    }
-    .mobile_mail {
-      display: block;
-    }
-  }
-  @include media-breakpoint-down(sm) {
-    font-size: 1.5rem;
-  }
-  &:after {
-    background: #3719ca;
-    content: '';
-    height: 2px;
-    position: absolute;
-    bottom: 0px;
-    left: 0;
-    width: 100%;
-    z-index: -1;
-    -webkit-transition: 0.2s;
-    -o-transition: 0.2s;
-    transition: 0.2s;
-    @include media-breakpoint-down(xs) {
-      content: none;
-    }
-  }
-  &:hover {
-    &:after {
-      height: 100%;
-    }
-    @include media-breakpoint-down(xs) {
-      color: $main_color;
-    }
-  }
-}
-.mobile_mail {
-  display: none;
-}
-
-.header-container {
-  display: flex;
-  justify-content: space-between;
-}
-
-.header-right {
-  display: flex;
-  align-items: center;
-}
-
-.burger-menu {
-  position: relative;
-  width: 3.4rem;
-  height: 3.4rem;
-  box-shadow: inset 0 0 0 1px #32344d;
-  border-radius: 50%;
-  transition: all 0.3s;
-  cursor: pointer;
-  margin-left: 2.86rem;
-  overflow: hidden;
-  transition: border-color 0.3s;
-  background-color: transparent;
-  border: none;
-
-  @media (min-width: 2500px) {
-    box-shadow: inset 0 0 0 2px #32344d;
-  }
-
-  .bar {
-    transition: all 0.3s;
-    height: 0.15rem;
-    width: 43%;
-    display: block;
-    background-color: #fff;
-    border-radius: 35px;
-    position: absolute;
-    top: 50%;
-    left: 0;
-    right: 0;
-    margin: auto;
-
-    &:before,
-    &:after {
-      left: 50%;
-      content: '';
-      display: block;
-      height: 0.15rem;
-      width: 70%;
-      background: #fff;
-      position: absolute;
-      border-radius: 35px;
-      transition: all 0.3s;
-      transform: translateX(-50%);
-    }
-    &:before {
-      margin-top: -0.5rem;
-    }
-    &:after {
-      margin-top: 0.5rem;
-    }
-  }
-
-  .titleHidden {
-    font-size: 0;
-  }
-
-  &.burger-menu--open {
-    .bar {
-      background-color: transparent;
-      &:before {
-        width: 100%;
-        top: 0.45rem;
-        transform: translateX(-50%) rotate(45deg);
-      }
-      &:after {
-        width: 100%;
-        transform: translateX(-50%) rotate(-45deg);
-        top: -0.55rem;
-      }
-    }
-  }
-}
-
-.burger-menu:after {
-  content: '';
-  position: absolute;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  display: block;
-  background: $main-color;
-  border-radius: 50%;
-  transition: transform 0.3s;
-  transform: translateX(100%);
-  z-index: -1;
-}
-
-.burger-menu:hover {
-  @include media-breakpoint-up(md) {
-    box-shadow: inset 0 0 0 1px $main-color;
-
-    &:after {
-      transform: translateX(0%);
-    }
-  }
-
-  @media (min-width: 2500px) {
-    box-shadow: inset 0 0 0 2px $main-color;
-  }
-}
-
-.header-left {
-  position: relative;
-  display: flex;
-  align-items: center;
-  margin-top: 2px;
-  z-index: 10;
-}
-.header-left a {
-  display: flex;
-}
-
-@include media-breakpoint-up(lg) {
-  .menu-item {
-    margin: 0;
-  }
-}
-
-.main-header .menu .menu-item {
-  opacity: 0;
-  @include media-breakpoint-up(lg) {
-    transform: translateY(50px);
-    transition: 0.5s;
-    transition-delay: 0;
-  }
-}
-
-.main-header--active .menu .menu-item__portfolio {
-  @include media-breakpoint-up(lg) {
-    transition-delay: 0.2s;
-  }
-}
-
-.main-header--active .menu .menu-item__services {
-  @include media-breakpoint-up(lg) {
-    transition-delay: 0.3s;
-  }
-}
-
-.main-header--active .menu .menu-item__blog {
-  @include media-breakpoint-up(lg) {
-    transition-delay: 0.4s;
-  }
-}
-
-.main-header--active .menu .menu-item__contacts {
-  @include media-breakpoint-up(lg) {
-    transition-delay: 0.5s;
-  }
-}
-
-.main-header--active .menu {
-  opacity: 1;
-}
-
-.main-header--active .mail {
-  opacity: 0;
-  pointer-events: none;
-}
-
-.main-header--active .main-menu {
-  opacity: 1;
-  pointer-events: auto;
-}
-
-.main-header--active .main-menu .menu-item {
-  opacity: 1;
-  @include media-breakpoint-up(lg) {
-    transform: translateY(0);
-  }
-}
