--- src/components/Header/components/Menu/Menu.scss
@@ -1,286 +1,0 @@
-@import '@styles/variables/all';
-@import '@styles/mixins/all';
-
-.main-menu {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  opacity: 0;
-  pointer-events: none;
-}
-.menu {
-  padding: 0;
-  margin: 0;
-  list-style: none;
-  display: flex;
-  width: 100%;
-  height: 100%;
-}
-.menu-item {
-  position: absolute;
-  margin: 0;
-
-  .active .menu-link__circle {
-    background-color: $grey-color;
-  }
-
-  @include media-breakpoint-down(xs) {
-    left: 50%;
-  }
-
-  a:before {
-    content: '';
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    pointer-events: none;
-    opacity: 0;
-    background-repeat: no-repeat;
-    background-position: center;
-    background-size: contain;
-    transition: opacity 0.5s;
-  }
-
-  a:after {
-    content: '';
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%) rotate(0deg);
-    pointer-events: none;
-    opacity: 0;
-    background-repeat: no-repeat;
-    background-position: center;
-    background-size: contain;
-    width: 19.86rem;
-    height: 19.86rem;
-    transform-origin: center;
-    transition: opacity 0.3s, transform 0.4s;
-  }
-
-  a {
-    display: flex;
-    align-items: center;
-    font-family: $font-sofia-pro;
-    font-size: 0.86rem;
-    font-weight: bold;
-    letter-spacing: 0.67em;
-    text-transform: uppercase;
-    color: #fafafa;
-    text-decoration: none;
-    transition: 0.2s;
-
-    @media (min-with: 2000px) {
-      font-size: 0.96rem;
-    }
-
-    @include media-breakpoint-down(sm) {
-      font-size: 1.167rem;
-      letter-spacing: 0.357em;
-    }
-  }
-}
-
-.menu-item:nth-child(odd) {
-  @include media-breakpoint-down(xs) {
-    transform: translateX(-185%);
-  }
-}
-
-.menu-item:nth-child(even) {
-  @include media-breakpoint-down(xs) {
-    transform: translateX(-33.33%);
-  }
-}
-
-.menu-item__blog {
-  top: 56.25%;
-
-  @include media-breakpoint-up(sm) {
-    left: 12.5%;
-  }
-
-  .menu-link__circle {
-    @include media-breakpoint-up(sm) {
-      width: 6.475rem;
-      height: 6.475rem;
-    }
-
-    @media (min-width: 2000px) {
-      width: 8.475rem;
-      height: 8.475rem;
-    }
-  }
-
-  a:before {
-    background-image: url('./images/blog-icon.svg');
-    width: 11.75rem;
-    height: 14rem;
-  }
-
-  a:after {
-    background-image: url('./images/blog-stars.svg');
-  }
-
-  &:hover .menu-link__circle {
-    @media (min-width: 1025px) {
-      transform: scale(3.07);
-    }
-  }
-}
-
-.menu-item__portfolio {
-  top: 21.375%;
-
-  @include media-breakpoint-up(sm) {
-    left: 30.39%;
-  }
-
-  .menu-link__circle {
-    @include media-breakpoint-up(sm) {
-      width: 10.24rem;
-      height: 10.24rem;
-    }
-
-    @media (min-width: 2000px) {
-      width: 12.24rem;
-      height: 12.24rem;
-    }
-  }
-
-  a:before {
-    background-image: url('./images/projects-icon.svg');
-    width: 11.33rem;
-    height: 11.33rem;
-  }
-
-  a:after {
-    background-image: url('./images/portfolio-stars.svg');
-  }
-
-  &:hover .menu-link__circle {
-    @media (min-width: 1025px) {
-      transform: scale(1.94);
-    }
-  }
-}
-
-.menu-item__services {
-  top: 39.125%;
-
-  @include media-breakpoint-up(sm) {
-    left: 69.6%;
-  }
-
-  .menu-link__circle {
-    @include media-breakpoint-up(sm) {
-      width: 5.858rem;
-      height: 5.858rem;
-    }
-
-    @media (min-width: 2000px) {
-      width: 7.858rem;
-      height: 7.858rem;
-    }
-  }
-
-  .menu-link__text {
-    @include media-breakpoint-up(sm) {
-      transform: translateX(-33%);
-    }
-  }
-
-  a:before {
-    background-image: url('./images/services-icon.svg');
-    width: 13.17rem;
-    height: 11.83rem;
-  }
-
-  a:after {
-    background-image: url('./images/services-stars.svg');
-  }
-
-  &:hover .menu-link__circle {
-    @media (min-width: 1025px) {
-      transform: scale(3.4);
-    }
-  }
-}
-
-.menu-item__contacts {
-  top: 71.75%;
-
-  @include media-breakpoint-up(sm) {
-    left: 53.9%;
-  }
-
-  .menu-link__circle {
-    @include media-breakpoint-up(sm) {
-      width: 7.75rem;
-      height: 7.75rem;
-    }
-
-    @media (min-width: 2000px) {
-      width: 9.858rem;
-      height: 9.858rem;
-    }
-  }
-
-  a:before {
-    background-image: url('./images/contacts-icon.svg');
-    width: 9.75rem;
-    height: 12.08rem;
-  }
-
-  a:after {
-    background-image: url('./images/contacts-stars.svg');
-  }
-
-  &:hover .menu-link__circle {
-    @media (min-width: 1025px) {
-      transform: scale(2.564);
-    }
-  }
-}
-
-.menu-link__circle {
-  border-radius: 50%;
-  background: $main-color;
-  transition: transform 0.5s;
-
-  @include media-breakpoint-down(xs) {
-    width: 5rem;
-    height: 5rem;
-  }
-}
-
-.menu-link__text {
-  position: absolute;
-  left: 100%;
-  transform: translateX(-50%);
-
-  @include media-breakpoint-down(xs) {
-    left: 50%;
-    transform: none;
-  }
-}
-
-.menu-item:hover {
-  @media (min-width: 1024px) {
-    a:before {
-      opacity: 1;
-      z-index: 1;
-      transition-delay: 0.2s;
-    }
-
-    a:after {
-      opacity: 1;
-      z-index: 1;
-      transform: translate(-50%, -50%) rotate(-20deg);
-      transition-delay: 0.2s;
-    }
-  }
-}
